feat(ngTranslate): Added translation into ngFormly
The form for editing the concours has now the correct labels. These are
now added via an expressionProperties that resturns a $translate
promise.

Also nested the concoursConstructor into the concoursListConstructor
with we $get in the provider of the module: models.concours.
That way the $translate is passed thru the my-parameter hash.