feat(ngInject): Intro of ng-annotate
Adds and removes AngularJS dependency injection annotations.  It removes
a lot of potential for mistakes in this area, like:
  o DI annotation missing
  o DI annotation incomplete
  o DI annotation in wrong order

Also used the ngStrictDi, the injector will be created in "strict-di"
mode. This means that the application will fail to invoke functions
which do not use explicit function annotation (and are thus unsuitable
for minification).